requirements.txt
asgiref==3.8.1
Brotli==1.1.0
dj-database-url==2.2.0
Django>=4.2
Django>=4.2
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
djangorestframework>=3.14.0
drf-yasg==1.21.7
gunicorn==23.0.0
inflection==0.5.1
packaging==24.1
psycopg2-binary==2.9.9
pytz==2024.2
PyYAML==6.0.2
setuptools==75.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
whitenoise==6.7.0
django-phonenumber-field==8.0.0
phonenumbers==8.13.47
